@@ -1,8 +1,5 @@
 package arrays;
 
-import java.util.Deque;
-import java.util.LinkedList;
-
 /**
  * Find intersection point, given that both arrays are sorted.
  * 
@@ -13,7 +10,7 @@ public class IntersectionPoint {
 	
 	public static void main(String[] s) {
 		int[] a = { 1, 2, 3, 6, 8, 10, 20 };
-		int[] b = { 0, 4, 5, 11, 15, 20 };
+		int[] b = { 0, 4, 8, 11, 15, 20 };
 		IntersectionPoint o = new IntersectionPoint();
 		System.out.println(o.find(a, b));
 	}
