@@ -3,55 +3,55 @@ package arrays;
 import java.util.Arrays;
 
 public class Problems {
-	
-	/**
-	 * Check palindrome
-	 */
-	static boolean checkPalindrom(int n) {
-		int divisor = 1;
-		while(n/divisor >= 10) {
-			divisor *= 10;
-		}
-		while(n > 0) {
-			int start = n/divisor;
-			int last = n%10;
-			if(start != last)
-				return false;
-			n = (n%divisor)/10;
-			divisor = divisor/100;
-		}
-		return true;
+
+	public static void main(String[] args) {
+		int[] a1 = {1,3};
+		int[] a2 = {2,4};
+		System.out.println(findMedianSortedArrays(a1, a2));
 	}
-	
-	/**
-	 * sort an array of 0, 1 and 2
-	 */
-	static void sortArrayOfZeroOneAndTwo(int[] a) {
-		int s = 0;
-		int mid = 0;
-		int e = a.length-1;
-		
-		while(mid <= e) {
-			if(a[mid] == 0) {
-				int t = a[mid];
-				a[mid] = a[s];
-				a[s] = t;
-				mid++; s++;
-			} else if(a[mid] == 1) {
-				mid++;
-			} else if(a[mid] == 2) {
-				int t = a[mid];
-				a[mid] = a[e];
-				a[e] = t;
-				e--;
+
+
+	public static double findMedianSortedArrays(int[] nums1, int[] nums2) {
+		if(nums1.length == 0 && nums2.length != 0)
+			return median(nums2);
+		if(nums1.length != 0 && nums2.length == 0)
+			return median(nums1);
+
+
+		int l1 = nums1.length;
+		int l2 = nums2.length;
+		int[] arr = new int[l1 + l2];
+		int pos = 0;
+		int p1 = 0;
+		int p2 = 0;
+		while(p1 < l1 && p2 < l2) {
+			if(nums1[p1] <= nums2[p2]) {
+				arr[pos++] = nums1[p1++];
+			} else {
+				arr[pos++] = nums2[p2++];
 			}
 		}
-		System.out.println(Arrays.toString(a));
+		if(p1 < l1) {
+			while(p1 < l1) {
+				arr[pos++] = nums1[p1++];
+			}
+		} else {
+			while(p2 < l2) {
+				arr[pos++] = nums2[p2++];
+			}
+		}
+		return median(arr);
 	}
-	
-	public static void main(String[] args) {
-		int [] a = {1,0,2,2,2,1,0,0,1,2,0,1};
-		System.out.println(checkPalindrom(14341));
+
+	private static double median(int[] arr) {
+		int l = arr.length;
+		int mid = l/2;
+		if(l % 2 == 0) {
+			return ((double) arr[mid] + (double) arr[mid-1])/2;
+		}
+		else {
+			return arr[mid];
+		}
 	}
 
 }
