@@ -1,25 +1,27 @@
 package linkedList;
 
-import java.util.Stack;
-
 public class ReverseLinkedList {
 
-	Node prev;
-	Node next;
+	private static Node prev;
+	private static Node next;
 	/**
 	 * In this method 2 pointers - next & prev are used.
 	 * we access every node one by one and move both pointers
 	 * in the end we assign prev to head/curr
 	 */
-	private Node reverserWithPointers(Node curr) {
-		while(curr != null) {
-			next = curr.next;
-			curr.next = prev;
-			prev = curr;
-			curr = next;
+	public static Node reverse(Node head) {
+		if(head == null || head.next == null)
+			return head;
+
+		while(head.next != null) {
+			next = head.next;
+			head.next = prev;
+			prev = head;
+			head = next;
 		}
-		curr = prev;
-		return curr;
+
+		head.next = prev;
+		return head;
 	}
 	
 	public static Node head;
@@ -36,19 +38,17 @@ public class ReverseLinkedList {
 		curr.next = null;
 		return head;
 	}
-	
+
 	public static void main(String[] args) {
 		Node x = new Node(1);
-		x.next = new Node(0);
-		x.next.next = new Node(0);
+		x.next = new Node(2);
+		x.next.next = new Node(3);
 		x.next.next.next = new Node(7);
 		
 		print(x);
-		ReverseLinkedList obj = new ReverseLinkedList();
-		head = x;
-		Node y = obj.reverseRecursive(head);
+		Node newHead = reverse(x);
 		System.out.println();
-		print(y);
+		print(newHead);
 	}
 	
 	public static void print(Node x) {
